reorganize tests dir
